added bio feature and remade some code

+++ functions/index.js
@@ -48,53 +48,49 @@ exports.deleteOldUser = functions.auth.user().onDelete(async (user) => {
 });
 
 exports.changeBanner = functions.region("us-central1").https.onCall(async (data, context) => {
-  // Takes in: auth, newBanner
+  // Verify that the function is called while authenticated
   if (!context.auth) {
-    // Throwing an HttpsError so that the client gets the error details.
-    throw new functions.https.HttpsError("failed-precondition", "The function must be called while authenticated.");
-  }
-  if (data.newBanner !== undefined && !validDataUrl(data.newBanner)) {
-    throw new functions.https.HttpsError("invalid-argument", "The function must be called with a valid data url.");
+    throw new functions.https.HttpsError("unauthenticated", "The function must be called while authenticated.");
   }
 
-  const [files] = await bucket.getFiles({ directory: `users/${context.auth.uid}/banner/` });
-  const userDoc = await admin.firestore().collection("users").doc(context.auth.uid).get();
+  // Verify that the new banner provided is a valid data URL
+  if (data.newBanner && !validDataUrl(data.newBanner)) {
+    throw new functions.https.HttpsError("invalid-argument", "The function must be called with a valid data URL.");
+  }
 
-  // Update the user's banner to changing
+  const userId = context.auth.uid;
+  const userRef = admin.firestore().collection("users").doc(userId);
+  let userDoc = await userRef.get();
+
+  // Update the user's banner to "changing"
   if (userDoc.exists) {
-    await admin.firestore().collection("users").doc(context.auth.uid).update({
+    await userRef.update({
       bannerChanging: true,
       bannerName: null,
     });
   } else {
-    await admin
-      .firestore()
-      .collection("users")
-      .doc(context.auth.uid)
-      .set({
-        displayName: (await admin.auth().getUser(context.auth.uid)).displayName,
-        uid: context.auth.uid,
-        bannerChanging: true,
-        bannerName: null,
-        avatarChanging: false,
-        avatarName: null,
-        creationTime: (await admin.auth().getUser(context.auth.uid)).metadata.creationTime,
-      });
+    // If the user does not exist in Firestore, create the document with default values
+    const { displayName, metadata } = await admin.auth().getUser(userId);
+    await userRef.set({
+      displayName: displayName,
+      uid: userId,
+      bannerChanging: true,
+      bannerName: null,
+      avatarChanging: false,
+      avatarName: null,
+      creationTime: metadata.creationTime,
+    });
   }
 
-  // Delete the old banner
-  await files.forEach(async (file) => {
-    await file.delete();
-  });
-
-  // Update the banner to what is provided
-  if (data.newBanner === undefined) {
-    // If no new banner is provided, update the banner to changed
-    await admin.firestore().collection("users").doc(context.auth.uid).update({
-      bannerChanging: false,
-      bannerName: null,
+  // Delete the old banner if it exists
+  let files = await bucket.getFiles({ directory: `users/${userId}/banner/` });
+  if (files[0].length > 0) {
+    await files[0].forEach(async (file) => {
+      await file.delete();
     });
-  } else {
+  }
+
+  if (data.newBanner) {
     // Generate new banner information
     const imageData = data.newBanner.replace(/^data:image\/\w+;base64,/, "");
     const imageBuffer = Buffer.from(imageData, "base64");
@@ -104,77 +100,69 @@ exports.changeBanner = functions.region("us-central1").https.onCall(async (data,
 
     // Upload the new banner
     let uploaded = false;
-    await bucket
-      .file(`users/${context.auth.uid}/banner/${uniqueName}`)
-      .save(imageBuffer, { public: true, gzip: true })
-      .then((data) => {
-        uploaded = true;
-      })
-      .catch(async (err) => {
-        console.log(err);
-      });
+    try {
+      await bucket.file(`users/${userId}/banner/${uniqueName}`).save(imageBuffer, { public: true, gzip: true });
+      uploaded = true;
+    } catch (err) {
+      console.error(err);
+    }
 
     // Update the user's banner information
-    await admin
-      .firestore()
-      .collection("users")
-      .doc(context.auth.uid)
-      .update({
-        bannerChanging: false,
-        bannerName: uploaded ? uniqueName : null,
-      });
+    await userRef.update({
+      bannerChanging: false,
+      bannerName: uploaded ? uniqueName : null,
+    });
+  } else {
+    // If no new banner is provided, update the banner to "not changing"
+    await userRef.update({ bannerChanging: false });
   }
   return "Complete";
 });
 
 exports.changeAvatar = functions.region("us-central1").https.onCall(async (data, context) => {
-  // Takes in: auth, newAvatar
+  // Verify that the function is called while authenticated
   if (!context.auth) {
-    // Throwing an HttpsError so that the client gets the error details.
-    throw new functions.https.HttpsError("failed-precondition", "The function must be called while authenticated.");
-  }
-  if (data.newAvatar !== undefined && !validDataUrl(data.newAvatar)) {
-    throw new functions.https.HttpsError("invalid-argument", "The function must be called with a valid data url.");
+    throw new functions.https.HttpsError("unauthenticated", "The function must be called while authenticated.");
   }
 
-  const [files] = await bucket.getFiles({ directory: `users/${context.auth.uid}/avatar/` });
-  const userDoc = await admin.firestore().collection("users").doc(context.auth.uid).get();
+  // Verify that the new avatar provided is a valid data URL
+  if (data.newAvatar && !validDataUrl(data.newAvatar)) {
+    throw new functions.https.HttpsError("invalid-argument", "The function must be called with a valid data URL.");
+  }
 
-  // Update the user's avatar to changing
+  const userId = context.auth.uid;
+  const userRef = admin.firestore().collection("users").doc(userId);
+  let userDoc = await userRef.get();
+
+  // Update the user's avatar to "changing"
   if (userDoc.exists) {
-    await admin.firestore().collection("users").doc(context.auth.uid).update({
+    await userRef.update({
       avatarChanging: true,
       avatarName: null,
     });
   } else {
-    await admin
-      .firestore()
-      .collection("users")
-      .doc(context.auth.uid)
-      .set({
-        displayName: (await admin.auth().getUser(context.auth.uid)).displayName,
-        uid: context.auth.uid,
-        bannerChanging: false,
-        bannerName: null,
-        avatarChanging: true,
-        avatarName: null,
-        creationTime: (await admin.auth().getUser(context.auth.uid)).metadata.creationTime,
-      });
+    // If the user does not exist in Firestore, create the document with default values
+    const { displayName, metadata } = await admin.auth().getUser(userId);
+    await userRef.set({
+      displayName: displayName,
+      uid: userId,
+      bannerChanging: false,
+      bannerName: null,
+      avatarChanging: true,
+      avatarName: null,
+      creationTime: metadata.creationTime,
+    });
   }
 
-  // Delete the old avatar
-  await files.forEach(async (file) => {
-    await file.delete();
-  });
-
-  // Update the avatar to what is provided
-  if (data.newAvatar === undefined) {
-    // If no new avatar is provided, update the avatar to changed
-    await admin.firestore().collection("users").doc(context.auth.uid).update({
-      avatarChanging: false,
-      avatarName: null,
+  // Delete the old avatar if it exists
+  let files = await bucket.getFiles({ directory: `users/${userId}/avatar/` });
+  if (files[0].length > 0) {
+    await files[0].forEach(async (file) => {
+      await file.delete();
     });
-  } else {
+  }
+
+  if (data.newAvatar) {
     // Generate new avatar information
     const imageData = data.newAvatar.replace(/^data:image\/\w+;base64,/, "");
     const imageBuffer = Buffer.from(imageData, "base64");
@@ -182,59 +170,97 @@ exports.changeAvatar = functions.region("us-central1").https.onCall(async (data,
     const imageHash = CryptoJS.MD5(CryptoJS.enc.Latin1.parse(data.newAvatar)).toString();
     const uniqueName = `${imageHash.substring(0, 16)}.${imageExtension}`;
 
-    // Upload the new banner
+    // Upload the new avatar
     let uploaded = false;
-    await bucket
-      .file(`users/${context.auth.uid}/avatar/${uniqueName}`)
-      .save(imageBuffer, { public: true, gzip: true })
-      .then((data) => {
-        uploaded = true;
-      })
-      .catch(async (err) => {
-        console.log(err);
-      });
+    try {
+      await bucket.file(`users/${userId}/avatar/${uniqueName}`).save(imageBuffer, { public: true, gzip: true });
+      uploaded = true;
+    } catch (err) {
+      console.error(err);
+    }
 
-    // Update the user's banner information
-    await admin
-      .firestore()
-      .collection("users")
-      .doc(context.auth.uid)
-      .update({
-        avatarChanging: false,
-        avatarName: uploaded ? uniqueName : null,
-      });
+    // Update the user's avatar information
+    await userRef.update({
+      avatarChanging: false,
+      avatarName: uploaded ? uniqueName : null,
+    });
+  } else {
+    // If no new avatar is provided, update the avatar to "not changing"
+    await userRef.update({ avatarChanging: false });
   }
   return "Complete";
 });
 
 exports.changeDisplayName = functions.region("us-central1").https.onCall(async (data, context) => {
-  // Takes in: auth, newBanner, hasBanner
+  // Verify that the function is called while authenticated
   if (!context.auth) {
-    // Throwing an HttpsError so that the client gets the error details.
-    throw new functions.https.HttpsError("failed-precondition", "The function must be called while authenticated.");
+    throw new functions.https.HttpsError("unauthenticated", "The function must be called while authenticated.");
   }
 
-  const userDoc = await admin.firestore().collection("users").doc(context.auth.uid).get();
+  // Verify that newDisplayName is provided
+  if (!data.newDisplayName) {
+    throw new functions.https.HttpsError("invalid-argument", "The function must be called with a new display name.");
+  }
 
-  // Update the user's banner to changing
+  const userId = context.auth.uid;
+  const userRef = admin.firestore().collection("users").doc(userId);
+  const userDoc = await userRef.get();
+
+  // Update the user's display name
   if (userDoc.exists) {
-    await admin.firestore().collection("users").doc(context.auth.uid).update({
+    await userRef.update({
       displayName: data.newDisplayName,
     });
   } else {
-    await admin
-      .firestore()
-      .collection("users")
-      .doc(context.auth.uid)
-      .set({
-        displayName: data.newDisplayName,
-        uid: context.auth.uid,
-        bannerChanging: false,
-        bannerName: null,
-        avatarChanging: false,
-        avatarName: null,
-        creationTime: (await admin.auth().getUser(context.auth.uid)).metadata.creationTime,
-      });
+    // If the user does not exist in Firestore, create the document with default values
+    const { metadata } = await admin.auth().getUser(userId);
+    await userRef.set({
+      displayName: data.newDisplayName,
+      uid: userId,
+      bannerChanging: false,
+      bannerName: null,
+      avatarChanging: false,
+      avatarName: null,
+      creationTime: metadata.creationTime,
+    });
   }
+
+  return "Complete";
+});
+
+exports.changeBio = functions.region("us-central1").https.onCall(async (data, context) => {
+  // Verify that the function is called while authenticated
+  if (!context.auth) {
+    throw new functions.https.HttpsError("unauthenticated", "The function must be called while authenticated.");
+  }
+
+  // Verify that newBio is provided
+  if (!data.newBio) {
+    throw new functions.https.HttpsError("invalid-argument", "The function must be called with a new bio.");
+  }
+
+  const userId = context.auth.uid;
+  const userRef = admin.firestore().collection("users").doc(userId);
+  const userDoc = await userRef.get();
+
+  // Update the user's bio
+  if (userDoc.exists) {
+    await userRef.update({
+      bio: data.newBio,
+    });
+  } else {
+    // If the user does not exist in Firestore, create the document with default values
+    const { metadata } = await admin.auth().getUser(userId);
+    await userRef.set({
+      bio: data.newBio,
+      uid: userId,
+      bannerChanging: false,
+      bannerName: null,
+      avatarChanging: false,
+      avatarName: null,
+      creationTime: metadata.creationTime,
+    });
+  }
+
   return "Complete";
 });
